Generate most of ParamPtr with a macro

+++ src/wrapper/vst3/wrapper.rs
@@ -336,7 +336,7 @@ impl<P: Vst3Plugin> IEditController for Wrapper<P> {
             u16strlcpy(&mut info.units, param_ptr.unit());
             // TODO: The host assumes these steps are distirbuted linearly, so this may cause weird
             //       behavior with skewed integers
-            info.step_count = param_ptr.step_count().unwrap_or(0);
+            info.step_count = param_ptr.step_count().unwrap_or(0) as i32;
             info.default_normalized_value = *default_value as f64;
             info.unit_id = vst3_sys::vst::kRootUnitId;
             info.flags = vst3_sys::vst::ParameterFlags::kCanAutomate as i32;
